+++ src/EvtGenModels/EvtModelReg.cpp
@@ -55,6 +55,7 @@
 #include "EvtGenModels/EvtBtoXsll.hh"
 #include "EvtGenModels/EvtCBTo3piMPP.hh"
 #include "EvtGenModels/EvtCBTo3piP00.hh"
+#include "EvtGenModels/EvtD0ToKspipi.hh"
 #include "EvtGenModels/EvtD0gammaDalitz.hh"
 #include "EvtGenModels/EvtD0mixDalitz.hh"
 #include "EvtGenModels/EvtDDalitz.hh"
@@ -273,6 +274,7 @@ EvtModelReg::EvtModelReg( const std::list<EvtDecayBase*>* extraModels )
     modelist.registerModel( new EvtDMix );
     modelist.registerModel( new EvtD0mixDalitz );
     modelist.registerModel( new EvtD0gammaDalitz );
+    modelist.registerModel( new EvtD0ToKspipi );
 
     modelist.registerModel( new EvtbTosllAli );
     modelist.registerModel( new EvtBaryonPCR );